+++ tilemap.cpp
@@ -220,7 +220,7 @@ bool TileMap::load(int level, int stage) {
 
     m_tileset.setSmooth(false);
 
-    int tilesPerRow = (m_tileset.getSize().x + 1) / (gTileSize + 1); // 1 pixel margin between tiles
+    int tilesPerRow = (m_tileset.getSize().x + 1) / (static_cast<int>(gTileSize) + 1); // Tiene en cuenta el pixel de margen entre tiles
 
     for (int i = 0; i < m_tilesPerRow; ++i)
     {
@@ -233,9 +233,9 @@ bool TileMap::load(int level, int stage) {
             int tileset_row = tileNumber % tilesPerRow;
             int tileset_column = tileNumber / tilesPerRow;
 
-            // Calculate tilemap coordinates (+1 for pixel margin between tiles)
-            int texX = tileset_row * (gTileSize + 1);
-            int texY = tileset_column * (gTileSize + 1);
+            // Calcular las coordenadas en el tilemap **saltando el píxel vacío**
+            int texX = tileset_row * (static_cast<int>(gTileSize) + 1); // Sumamos el 1px de separación
+            int texY = tileset_column * (static_cast<int>(gTileSize) + 1);
 
             // Points to the first vertex of the tile
             sf::Vertex* triangle = &m_vertices[tileIndex * 6];
